Move Snapshot type tests along the jest tests for snapshot. Rewrite these tests as noot jest tests.
